+++ family_foto/utils/thumbnail_service.py
