+++ tests/routes/test_resized.py
@@ -1,6 +1,6 @@
 from flask_api import status
 
-from family_foto.utils.ThumbnailService import ThumbnailService
+from family_foto.utils.thumbnail_service import ThumbnailService
 from tests.base_login_test_case import BaseLoginTestCase
 from tests.base_photo_test_case import BasePhotoTestCase
