+++ tests/utils/test_thumbnail_service.py
@@ -1,5 +1,5 @@
 from family_foto.models.photo import Photo
-from family_foto.utils.ThumbnailService import ThumbnailService
+from family_foto.utils.thumbnail_service import ThumbnailService
 from tests.base_photo_test_case import BasePhotoTestCase
